requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.4
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.2.2
python-decouple==3.8
requests==2.32.3
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
